backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.22.0
motor==3.1.1
pydantic==1.10.8
openai==0.27.8
python-dotenv==1.0.0
